Refactor sequential interface to use new call mechanism
- Replaced instances of `sequential_dispatch.h` with `call.h` in multiple sequential interface headers to streamline the dispatch logic.
- Updated function calls to utilize the new `cpp_core::internal::sequential::call` method, ensuring consistent handling of sequential execution across the interface.
- Improved formatting in several function implementations for better readability.

These changes enhance the maintainability and clarity of the serial interface, aligning with recent refactoring efforts.
