+++ include/cpp_core/interface/sequential/serial_clear_buffer_in_sequential.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../../error_callback.h"
-#include "../../internal/sequential_dispatch.h"
+#include "../../internal/sequential/call.h"
 #include "../serial_clear_buffer_in.h"
 
 #ifdef __cplusplus
@@ -18,7 +18,9 @@ extern "C"
         ErrorCallbackT error_callback = nullptr
     ) -> int
     {
-        return cpp_core::internal::seq::call(handle, [=] { return serialClearBufferIn(handle, error_callback); });
+        return cpp_core::internal::sequential::call(handle, [=] {
+            return serialClearBufferIn(handle, error_callback);
+        });
     }
 
 #ifdef __cplusplus
