+++ include/cpp_core/interface/sequential/serial_read_until_sequential.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../../error_callback.h"
-#include "../../internal/sequential_dispatch.h"
+#include "../../internal/sequential/call.h"
 #include "../serial_read_until.h"
 
 #ifdef __cplusplus
@@ -23,7 +23,7 @@ extern "C"
         ErrorCallbackT error_callback = nullptr
     ) -> int
     {
-        return cpp_core::internal::seq::call(handle, [=] {
+        return cpp_core::internal::sequential::call(handle, [=] {
             return serialReadUntil(handle, buffer, buffer_size, timeout_ms, multiplier, until_char, error_callback);
         });
     }
